FFPRestful/requirements.txt
django==1.11.3
pillow==4.3.0
mysqlclient==1.3.12
djangorestframework-jwt==1.11.0
djangorestframework==3.7
markdown==2.6.10
django-filter==1.1.0
django-crispy-forms==1.7.0
django-reversion==2.0.12
django-formtools==2.1
future==0.16.0
httplib2==0.10.3
six==1.11.0
requests==2.18.4
coreapi==2.3.3
django-cors-headers==2.1.0
xlwt==1.3.0
xlsxwriter==1.0.2
drf-extensions==0.3.1